Undo breaking API change

+++ backoff.go
@@ -27,7 +27,7 @@ type Backoff struct {
 // Duration returns the duration for the current attempt before incrementing
 // the attempt counter. See ForAttempt.
 func (b *Backoff) Duration() time.Duration {
-	d := b.ForAttempt(atomic.LoadUint64(&b.attempt))
+	d := b.ForAttempt(float64(atomic.LoadUint64(&b.attempt)))
 	atomic.AddUint64(&b.attempt, 1)
 	return d
 }
@@ -40,7 +40,7 @@ const maxInt64 = float64(math.MaxInt64 - 512)
 // attempt should be 0.
 //
 // ForAttempt is concurrent-safe.
-func (b *Backoff) ForAttempt(attempt uint64) time.Duration {
+func (b *Backoff) ForAttempt(attempt float64) time.Duration {
 	// Zero-values are nonsensical, so we use
 	// them to apply defaults
 	min := b.Min
@@ -61,7 +61,7 @@ func (b *Backoff) ForAttempt(attempt uint64) time.Duration {
 	}
 	//calculate this duration
 	minf := float64(min)
-	durf := minf * math.Pow(factor, float64(attempt))
+	durf := minf * math.Pow(factor, attempt)
 	if b.Jitter {
 		durf = rand.Float64()*(durf-minf) + minf
 	}
@@ -86,8 +86,8 @@ func (b *Backoff) Reset() {
 }
 
 // Attempt returns the current attempt counter value.
-func (b *Backoff) Attempt() uint64 {
-	return atomic.LoadUint64(&b.attempt)
+func (b *Backoff) Attempt() float64 {
+	return float64(atomic.LoadUint64(&b.attempt))
 }
 
 // Copy returns a backoff with equals constraints as the original
