+++ backoff_test.go
@@ -83,17 +83,17 @@ func TestGetAttempt(t *testing.T) {
 		Max:    10 * time.Second,
 		Factor: 2,
 	}
-	equals(t, b.Attempt(), uint64(0))
+	equals(t, b.Attempt(), float64(0))
 	equals(t, b.Duration(), 100*time.Millisecond)
-	equals(t, b.Attempt(), uint64(1))
+	equals(t, b.Attempt(), float64(1))
 	equals(t, b.Duration(), 200*time.Millisecond)
-	equals(t, b.Attempt(), uint64(2))
+	equals(t, b.Attempt(), float64(2))
 	equals(t, b.Duration(), 400*time.Millisecond)
-	equals(t, b.Attempt(), uint64(3))
+	equals(t, b.Attempt(), float64(3))
 	b.Reset()
-	equals(t, b.Attempt(), uint64(0))
+	equals(t, b.Attempt(), float64(0))
 	equals(t, b.Duration(), 100*time.Millisecond)
-	equals(t, b.Attempt(), uint64(1))
+	equals(t, b.Attempt(), float64(1))
 }
 
 func TestJitter(t *testing.T) {
